Explorer.JSONRPC -> EthereumJSONRPC
Extract Explorer.JSONRPC to its own OTP application,
apps/ethereum_jsonrpc.  It is not ready to be a completely separate
library because it knowns the internal params format used for
Explorer.Chain.
