+++ apps/ethereum_jsonrpc/README.md
@@ -1,21 +1,36 @@
-# EthereumJsonrpc
+# EthereumJSONRPC
 
-**TODO: Add description**
+Ethereum JSONRPC client.
+
+## Configuration
+
+Configuration for parity URLs can be provided with the following mix
+config:
+
+```elixir
+config :ethereum_jsonrpc,
+  url: "https://sokol.poa.network",
+  trace_url: "https://sokol-trace.poa.network",
+  http: [recv_timeout: 60_000, timeout: 60_000, hackney: [pool: :ethereum_jsonrpc]]
+```
+
+Note: the tracing node URL is provided separately from `:url`,
+via `:trace_url`. The trace URL and is used for
+`fetch_internal_transactions`, which is only a supported method on
+tracing nodes. The `:http` option is passed directly to the HTTP
+library (`HTTPoison`), which forwards the options down to `:hackney`.
 
 ## Installation
 
-If [available in Hex](https://hex.pm/docs/publish), the package can be installed
-by adding `ethereum_jsonrpc` to your list of dependencies in `mix.exs`:
+The OTP application `:ethereum_jsonrpc` can be used in other umbrella
+OTP applications by adding `ethereum_jsonrpc` to your list of
+dependencies in `mix.exs`:
 
 ```elixir
 def deps do
   [
-    {:ethereum_jsonrpc, "~> 0.1.0"}
+    {:ethereum_jsonrpc, in_umbrella: true}
   ]
 end
 ```
-
-Documentation can be generated with [ExDoc](https://github.com/elixir-lang/ex_doc)
-and published on [HexDocs](https://hexdocs.pm). Once published, the docs can
-be found at [https://hexdocs.pm/ethereum_jsonrpc](https://hexdocs.pm/ethereum_jsonrpc).
