+++ apps/ethereum_jsonrpc/lib/ethereum_jsonrpc/block.ex
@@ -1,55 +1,55 @@
-defmodule Explorer.JSONRPC.Block do
+defmodule EthereumJSONRPC.Block do
   @moduledoc """
   Block format as returned by [`eth_getBlockByHash`](https://github.com/ethereum/wiki/wiki/JSON-RPC#eth_getblockbyhash)
   and [`eth_getBlockByNumber`](https://github.com/ethereum/wiki/wiki/JSON-RPC#eth_getblockbynumber).
   """
 
-  import Explorer.JSONRPC, only: [nonce_to_integer: 1, quantity_to_integer: 1, timestamp_to_datetime: 1]
+  import EthereumJSONRPC, only: [nonce_to_integer: 1, quantity_to_integer: 1, timestamp_to_datetime: 1]
 
-  alias Explorer.JSONRPC
-  alias Explorer.JSONRPC.Transactions
+  alias EthereumJSONRPC
+  alias EthereumJSONRPC.Transactions
 
   @type elixir :: %{String.t() => non_neg_integer | DateTime.t() | String.t() | nil}
 
   @typedoc """
-   * `"author"` - `t:Explorer.JSONRPC.address/0` that created the block.  Aliased by `"miner"`.
-   * `"difficulty"` - `t:Explorer.JSONRPC.quantity/0`` of the difficulty for this block.
-   * `"extraData"` - the extra `t:Explorer.JSONRPC.data/0`` field of this block.
-   * `"gasLimit" - maximum gas `t:Explorer.JSONRPC.quantity/0`` in this block.
-   * `"gasUsed" - the total `t:Explorer.JSONRPC.quantity/0`` of gas used by all transactions in this block.
-   * `"hash"` - the `t:Explorer.JSONRPC.hash/0` of the block.
-   * `"logsBloom"` - `t:Explorer.JSONRPC.data/0`` for the [Bloom filter](https://en.wikipedia.org/wiki/Bloom_filter)
+   * `"author"` - `t:EthereumJSONRPC.address/0` that created the block.  Aliased by `"miner"`.
+   * `"difficulty"` - `t:EthereumJSONRPC.quantity/0`` of the difficulty for this block.
+   * `"extraData"` - the extra `t:EthereumJSONRPC.data/0`` field of this block.
+   * `"gasLimit" - maximum gas `t:EthereumJSONRPC.quantity/0`` in this block.
+   * `"gasUsed" - the total `t:EthereumJSONRPC.quantity/0`` of gas used by all transactions in this block.
+   * `"hash"` - the `t:EthereumJSONRPC.hash/0` of the block.
+   * `"logsBloom"` - `t:EthereumJSONRPC.data/0`` for the [Bloom filter](https://en.wikipedia.org/wiki/Bloom_filter)
      for the logs of the block. `nil` when block is pending.
-   * `"miner"` - `t:Explorer.JSONRPC.address/0` of the beneficiary to whom the mining rewards were given.  Aliased by
+   * `"miner"` - `t:EthereumJSONRPC.address/0` of the beneficiary to whom the mining rewards were given.  Aliased by
       `"author"`.
-   * `"nonce"` -  `t:Explorer.JSONRPC.nonce/0`. `nil` when its pending block.
-   * `"number"` - the block number `t:Explorer.JSONRPC.quantity/0`. `nil` when block is pending.
-   * `"parentHash" - the `t:Explorer.JSONRPC.hash/0` of the parent block.
-   * `"receiptsRoot"` - `t:Explorer.JSONRPC.hash/0` of the root of the receipts.
+   * `"nonce"` -  `t:EthereumJSONRPC.nonce/0`. `nil` when its pending block.
+   * `"number"` - the block number `t:EthereumJSONRPC.quantity/0`. `nil` when block is pending.
+   * `"parentHash" - the `t:EthereumJSONRPC.hash/0` of the parent block.
+   * `"receiptsRoot"` - `t:EthereumJSONRPC.hash/0` of the root of the receipts.
      [trie](https://github.com/ethereum/wiki/wiki/Patricia-Tree) of the block.
    * `"sealFields"` - UNKNOWN
-   * `"sha3Uncles"` - `t:Explorer.JSONRPC.hash/0` of the
+   * `"sha3Uncles"` - `t:EthereumJSONRPC.hash/0` of the
      [uncles](https://bitcoin.stackexchange.com/questions/39329/in-ethereum-what-is-an-uncle-block) data in the block.
    * `"signature"` - UNKNOWN
-   * `"size"` - `t:Explorer.JSONRPC.quantity/0`` of bytes in this block
-   * `"stateRoot" - `t:Explorer.JSONRPC.hash/0` of the root of the final state
+   * `"size"` - `t:EthereumJSONRPC.quantity/0`` of bytes in this block
+   * `"stateRoot" - `t:EthereumJSONRPC.hash/0` of the root of the final state
      [trie](https://github.com/ethereum/wiki/wiki/Patricia-Tree) of the block.
    * `"step"` - UNKNOWN
-   * `"timestamp"`: the unix timestamp as a `t:Explorer.JSONRPC.quantity/0`` for when the block was collated.
-   * `"totalDifficulty" - `t:Explorer.JSONRPC.quantity/0`` of the total difficulty of the chain until this block.
-   * `"transactions"` - `t:list/0` of `t:Explorer.JSONRPC.Transaction.t/0`.
-   * `"transactionsRoot" - `t:Explorer.JSONRPC.hash/0` of the root of the transaction
+   * `"timestamp"`: the unix timestamp as a `t:EthereumJSONRPC.quantity/0`` for when the block was collated.
+   * `"totalDifficulty" - `t:EthereumJSONRPC.quantity/0`` of the total difficulty of the chain until this block.
+   * `"transactions"` - `t:list/0` of `t:EthereumJSONRPC.Transaction.t/0`.
+   * `"transactionsRoot" - `t:EthereumJSONRPC.hash/0` of the root of the transaction
      [trie](https://github.com/ethereum/wiki/wiki/Patricia-Tree) of the block.
    * `uncles`: `t:list/0` of
      [uncles](https://bitcoin.stackexchange.com/questions/39329/in-ethereum-what-is-an-uncle-block)
-     `t:Explorer.JSONRPC.hash/0`.
+     `t:EthereumJSONRPC.hash/0`.
   """
-  @type t :: %{String.t() => JSONRPC.data() | JSONRPC.hash() | JSONRPC.quantity() | nil}
+  @type t :: %{String.t() => EthereumJSONRPC.data() | EthereumJSONRPC.hash() | EthereumJSONRPC.quantity() | nil}
 
   @doc """
   Converts `t:elixir/0` format to params used in `Explorer.Chain`.
 
-      iex> Explorer.JSONRPC.Block.elixir_to_params(
+      iex> EthereumJSONRPC.Block.elixir_to_params(
       ...>   %{
       ...>     "author" => "0xe8ddc5c7a2d2f0d7a9798459c0104fdf5e987aca",
       ...>     "difficulty" => 340282366920938463463374607431465537093,
@@ -125,9 +125,9 @@ defmodule Explorer.JSONRPC.Block do
   end
 
   @doc """
-  Get `t:Explorer.JSONRPC.Transactions.elixir/0` from `t:elixir/0`
+  Get `t:EthereumJSONRPC.Transactions.elixir/0` from `t:elixir/0`
 
-      iex> Explorer.JSONRPC.Block.elixir_to_transactions(
+      iex> EthereumJSONRPC.Block.elixir_to_transactions(
       ...>   %{
       ...>     "author" => "0xe8ddc5c7a2d2f0d7a9798459c0104fdf5e987aca",
       ...>     "difficulty" => 340282366920938463463374607431768211454,
@@ -211,7 +211,7 @@ defmodule Explorer.JSONRPC.Block do
   @doc """
   Decodes the stringly typed numerical fields to `t:non_neg_integer/0` and the timestamps to `t:DateTime.t/0`
 
-      iex> Explorer.JSONRPC.Block.to_elixir(
+      iex> EthereumJSONRPC.Block.to_elixir(
       ...>   %{
       ...>     "author" => "0xe8ddc5c7a2d2f0d7a9798459c0104fdf5e987aca",
       ...>     "difficulty" => "0xfffffffffffffffffffffffffffffffe",
@@ -278,7 +278,7 @@ defmodule Explorer.JSONRPC.Block do
   end
 
   # double check that no new keys are being missed by requiring explicit match for passthrough
-  # `t:Explorer.JSONRPC.address/0` and `t:Explorer.JSONRPC.hash/0` pass through as `Explorer.Chain` can verify correct
+  # `t:EthereumJSONRPC.address/0` and `t:EthereumJSONRPC.hash/0` pass through as `Explorer.Chain` can verify correct
   # hash format
   defp entry_to_elixir({key, _} = entry)
        when key in ~w(author extraData hash logsBloom miner parentHash receiptsRoot sealFields sha3Uncles signature
