+++ apps/ethereum_jsonrpc/lib/ethereum_jsonrpc/parity/trace/result.ex
@@ -1,14 +1,14 @@
-defmodule Explorer.JSONRPC.Parity.Trace.Result do
+defmodule EthereumJSONRPC.Parity.Trace.Result do
   @moduledoc """
-  The result of performing the `t:Explorer.JSONRPC.Parity.Action.t/0` in a `t:Explorer.JSONRPC.Parity.Trace.t/0`.
+  The result of performing the `t:EthereumJSONRPC.Parity.Action.t/0` in a `t:EthereumJSONRPC.Parity.Trace.t/0`.
   """
 
-  import Explorer.JSONRPC, only: [quantity_to_integer: 1]
+  import EthereumJSONRPC, only: [quantity_to_integer: 1]
 
   @doc """
   Decodes the stringly typed numerical fields to `t:non_neg_integer/0`.
 
-      iex> Explorer.JSONRPC.Parity.Trace.Result.to_elixir(
+      iex> EthereumJSONRPC.Parity.Trace.Result.to_elixir(
       ...>   %{
       ...>     "address" => "0xffc87239eb0267bc3ca2cd51d12fbf278e02ccb4",
       ...>     "code" => "0x606060405260043610610062576000357c0100000000000000000000000000000000000000000000000000000000900463ffffffff1680630900f01014610067578063445df0ac146100a05780638da5cb5b146100c9578063fdacd5761461011e575b600080fd5b341561007257600080fd5b61009e600480803573ffffffffffffffffffffffffffffffffffffffff16906020019091905050610141565b005b34156100ab57600080fd5b6100b3610224565b6040518082815260200191505060405180910390f35b34156100d457600080fd5b6100dc61022a565b604051808273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200191505060405180910390f35b341561012957600080fd5b61013f600480803590602001909190505061024f565b005b60008060009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff163373ffffffffffffffffffffffffffffffffffffffff161415610220578190508073ffffffffffffffffffffffffffffffffffffffff1663fdacd5766001546040518263ffffffff167c010000000000000000000000000000000000000000000000000000000002815260040180828152602001915050600060405180830381600087803b151561020b57600080fd5b6102c65a03f1151561021c57600080fd5b5050505b5050565b60015481565b6000809054906101000a900473ffffffffffffffffffffffffffffffffffffffff1681565b6000809054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff163373ffffffffffffffffffffffffffffffffffffffff1614156102ac57806001819055505b505600a165627a7a72305820a9c628775efbfbc17477a472413c01ee9b33881f550c59d21bee9928835c854b0029",
@@ -23,7 +23,7 @@ defmodule Explorer.JSONRPC.Parity.Trace.Result do
 
   `nil` resultscan occur for suicide type traces.
 
-      iex> Explorer.JSONRPC.Parity.Trace.Result.to_elixir(nil)
+      iex> EthereumJSONRPC.Parity.Trace.Result.to_elixir(nil)
       nil
 
   """
