+++ apps/ethereum_jsonrpc/mix.exs
@@ -3,31 +3,63 @@ defmodule EthereumJsonrpc.MixProject do
 
   def project do
     [
+      aliases: aliases(Mix.env()),
       app: :ethereum_jsonrpc,
-      version: "0.1.0",
       build_path: "../../_build",
       config_path: "../../config/config.exs",
+      deps: deps(),
       deps_path: "../../deps",
-      lockfile: "../../mix.lock",
+      dialyzer: [
+        plt_add_deps: :transitive,
+        plt_add_apps: [:mix],
+        ignore_warnings: "../../.dialyzer-ignore"
+      ],
       elixir: "~> 1.6",
+      lockfile: "../../mix.lock",
+      preferred_cli_env: [
+        coveralls: :test,
+        "coveralls.detail": :test,
+        "coveralls.post": :test,
+        "coveralls.html": :test,
+        dialyzer: :test
+      ],
       start_permanent: Mix.env() == :prod,
-      deps: deps()
+      test_coverage: [tool: ExCoveralls],
+      version: "0.1.0"
     ]
   end
 
   # Run "mix help compile.app" to learn about applications.
   def application do
     [
+      mod: {EthereumJSONRPC.Application, []},
       extra_applications: [:logger]
     ]
   end
 
+  defp aliases(env) do
+    env_aliases(env)
+  end
+
+  defp env_aliases(:dev), do: []
+
+  defp env_aliases(_env), do: [compile: "compile --warnings-as-errors"]
+
   # Run "mix help deps" to learn about dependencies.
   defp deps do
     [
-      # {:dep_from_hexpm, "~> 0.3.0"},
-      # {:dep_from_git, git: "https://github.com/elixir-lang/my_dep.git", tag: "0.1.0"},
-      # {:sibling_app_in_umbrella, in_umbrella: true},
+      # Style Checking
+      {:credo, "0.9.2", only: [:dev, :test], runtime: false},
+      # Static Type Checking
+      {:dialyxir, "~> 0.5", only: [:dev, :test], runtime: false},
+      # Code coverage
+      {:excoveralls, "~> 0.8.1", only: [:test]},
+      # JSONRPC HTTP Post calls
+      {:httpoison, "~> 1.0", override: true},
+      # Decode/Encode JSON for JSONRPC
+      {:jason, "~> 1.0"},
+      # Convert unix timestamps in JSONRPC to DateTimes
+      {:timex, "~> 3.1.24"}
     ]
   end
 end
