+++ apps/ethereum_jsonrpc/test/ethereum_jsonrpc/block_test.exs
@@ -1,0 +1,5 @@
+defmodule EthereumJSONRPC.BlockTest do
+  use ExUnit.Case, async: true
+
+  doctest EthereumJSONRPC.Block
+end
