+++ apps/ethereum_jsonrpc/test/ethereum_jsonrpc/blocks_test.exs
@@ -1,0 +1,5 @@
+defmodule EthereumJSONRPC.BlocksTest do
+  use ExUnit.Case, async: true
+
+  doctest EthereumJSONRPC.Blocks
+end
