+++ apps/ethereum_jsonrpc/test/ethereum_jsonrpc/log_test.exs
@@ -1,0 +1,5 @@
+defmodule EthereumJSONRPC.LogTest do
+  use ExUnit.Case, async: true
+
+  doctest EthereumJSONRPC.Log
+end
