+++ apps/ethereum_jsonrpc/test/ethereum_jsonrpc/parity/trace/action_test.exs
@@ -1,0 +1,5 @@
+defmodule EthereumJSONRPC.Parity.Trace.ActionTest do
+  use ExUnit.Case, async: true
+
+  doctest EthereumJSONRPC.Parity.Trace.Action
+end
