+++ apps/ethereum_jsonrpc/test/ethereum_jsonrpc/parity/trace/result_test.exs
@@ -1,0 +1,5 @@
+defmodule EthereumJSONRPC.Parity.Trace.ResultTest do
+  use ExUnit.Case, async: true
+
+  doctest EthereumJSONRPC.Parity.Trace.Result
+end
