+++ apps/ethereum_jsonrpc/test/ethereum_jsonrpc/parity/trace_test.exs
@@ -1,0 +1,5 @@
+defmodule EthereumJSONRPC.Parity.TraceTest do
+  use ExUnit.Case, async: true
+
+  doctest EthereumJSONRPC.Parity.Trace
+end
