+++ apps/ethereum_jsonrpc/test/ethereum_jsonrpc/parity_test.exs
@@ -1,0 +1,5 @@
+defmodule EthereumJSONRPC.ParityTest do
+  use ExUnit.Case, async: true
+
+  doctest EthereumJSONRPC.Parity
+end
