+++ apps/ethereum_jsonrpc/test/ethereum_jsonrpc/receipt_test.exs
@@ -1,0 +1,5 @@
+defmodule EthereumJSONRPC.ReceiptTest do
+  use ExUnit.Case, async: true
+
+  doctest EthereumJSONRPC.Receipt
+end
