+++ apps/ethereum_jsonrpc/test/ethereum_jsonrpc/transaction_test.exs
@@ -1,0 +1,5 @@
+defmodule EthereumJSONRPC.TransactionTest do
+  use ExUnit.Case, async: true
+
+  doctest EthereumJSONRPC.Transaction
+end
