+++ apps/ethereum_jsonrpc/test/ethereum_jsonrpc/transactions_test.exs
@@ -1,0 +1,5 @@
+defmodule EthereumJSONRPC.TransactionsTest do
+  use ExUnit.Case, async: true
+
+  doctest EthereumJSONRPC.Transactions
+end
