--- apps/ethereum_jsonrpc/test/ethereum_jsonrpc_test.exs
@@ -1,8 +1,0 @@
-defmodule EthereumJsonrpcTest do
-  use ExUnit.Case
-  doctest EthereumJsonrpc
-
-  test "greets the world" do
-    assert EthereumJsonrpc.hello() == :world
-  end
-end
