+++ apps/explorer/mix.exs
@@ -70,6 +70,8 @@ defmodule Explorer.Mixfile do
       # Code coverage
       {:excoveralls, "~> 0.8.1", only: [:test]},
       {:exvcr, "~> 0.10", only: :test},
+      # JSONRPC access to Parity for `Explorer.Indexer`
+      {:ethereum_jsonrpc, in_umbrella: true},
       {:httpoison, "~> 1.0", override: true},
       {:jason, "~> 1.0"},
       {:junit_formatter, ">= 0.0.0", only: [:test], runtime: false},
