--- apps/explorer/test/explorer/jsonrpc/block_test.exs
@@ -1,5 +1,0 @@
-defmodule Explorer.JSONRPC.BlockTest do
-  use ExUnit.Case, async: true
-
-  doctest Explorer.JSONRPC.Block
-end
