--- apps/explorer/test/explorer/jsonrpc/blocks_test.exs
@@ -1,5 +1,0 @@
-defmodule Explorer.JSONRPC.BlocksTest do
-  use ExUnit.Case, async: true
-
-  doctest Explorer.JSONRPC.Blocks
-end
