--- apps/explorer/test/explorer/jsonrpc/log_test.exs
@@ -1,5 +1,0 @@
-defmodule Explorer.JSONRPC.LogTest do
-  use ExUnit.Case, async: true
-
-  doctest Explorer.JSONRPC.Log
-end
