--- apps/explorer/test/explorer/jsonrpc/parity/trace/action_test.exs
@@ -1,5 +1,0 @@
-defmodule Explorer.JSONRPC.Parity.Trace.ActionTest do
-  use ExUnit.Case, async: true
-
-  doctest Explorer.JSONRPC.Parity.Trace.Action
-end
