--- apps/explorer/test/explorer/jsonrpc/parity/trace/result_test.exs
@@ -1,5 +1,0 @@
-defmodule Explorer.JSONRPC.Parity.Trace.ResultTest do
-  use ExUnit.Case, async: true
-
-  doctest Explorer.JSONRPC.Parity.Trace.Result
-end
