--- apps/explorer/test/explorer/jsonrpc/parity/trace_test.exs
@@ -1,5 +1,0 @@
-defmodule Explorer.JSONRPC.Parity.TraceTest do
-  use ExUnit.Case, async: true
-
-  doctest Explorer.JSONRPC.Parity.Trace
-end
