--- apps/explorer/test/explorer/jsonrpc/parity_test.exs
@@ -1,5 +1,0 @@
-defmodule Explorer.JSONRPC.ParityTest do
-  use ExUnit.Case, async: true
-
-  doctest Explorer.JSONRPC.Parity
-end
