--- apps/explorer/test/explorer/jsonrpc/receipt_test.exs
@@ -1,5 +1,0 @@
-defmodule Explorer.JSONRPC.ReceiptTest do
-  use ExUnit.Case, async: true
-
-  doctest Explorer.JSONRPC.Receipt
-end
