--- apps/explorer/test/explorer/jsonrpc/transaction_test.exs
@@ -1,5 +1,0 @@
-defmodule Explorer.JSONRPC.TransactionTest do
-  use ExUnit.Case, async: true
-
-  doctest Explorer.JSONRPC.Transaction
-end
