--- apps/explorer/test/explorer/jsonrpc/transactions_test.exs
@@ -1,5 +1,0 @@
-defmodule Explorer.JSONRPC.TransactionsTest do
-  use ExUnit.Case, async: true
-
-  doctest Explorer.JSONRPC.Transactions
-end
