+++ coveralls.json
@@ -1,7 +1,7 @@
 {
   "coverage_options": {
     "treat_no_relevant_lines_as_covered": true,
-    "minimum_coverage": 94.0
+    "minimum_coverage": 91.2
   },
   "terminal_options": {
     "file_column_width": 120
